# Add upper limit for the number of completion items