I'm using rust-analyzer on a project with a `bindgen`'d file that contains a large number of functions and structs. Code completion on this project took extremely long due to the large number of items that are rendered and sent to the IDE.

This PR drastically improves performance in cases where lots of completion results are found by adding an upper limit to the list of completion items. Comparison on a (real-world) extreme case with over 200,000 items in one crate:

# Before
```
13557ms - handle_completion
       18ms - CompletionContext::new
           14ms - CompletionContext::expand_and_fill
                0   - Semantics::analyze_impl (1 calls)
                0   - SourceBinder::to_module_def (4 calls)
                0   - crate_def_map:wait (2 calls)
               14ms - ???
            0   - Semantics::analyze_impl (1 calls)
            0   - crate_def_map:wait (2 calls)
            1ms - descend_into_macros (1 calls)
    11750ms - complete_expr_path
         1226ms - render_resolution
             1226ms - render_fn
                 1226ms - lang_item_query
                     1222ms - crate_lang_items_query
                            6ms - collect_lang_item
                           18ms - collect_lang_item
                            9ms - collect_lang_item
                         1155ms - collect_lang_item (344777 calls)
                            0   - crate_def_map:wait (1 calls)
                           31ms - ???
                    0   - crate_def_map:wait (6 calls)
           11ms - render_resolution
               11ms - render_fn
                    0   - crate_def_map:wait (6 calls)
                   11ms - ???
           22ms - PerNs::filter_visibility (235244 calls)
          160ms - crate_def_map:wait (235245 calls)
          781ms - generic_params_query (133796 calls)
           88ms - item::Builder::build (235244 calls)
         6772ms - render_resolution (235242 calls)
         2687ms - ???
        0   - Semantics::analyze_impl (1 calls)
        0   - crate_def_map:wait (2 calls)
     1788ms - ???
```

In practise, it takes much longer for the code completion result to show up in the IDE since processing such a large list puts quite a burden on VS code as well (and/or it's an IO bottleneck during communication between the extension and rust-analyzer).

# After
```
1988ms - handle_completion
       18ms - CompletionContext::new
           15ms - CompletionContext::expand_and_fill
                0   - Semantics::analyze_impl (1 calls)
                0   - SourceBinder::to_module_def (4 calls)
                0   - crate_def_map:wait (2 calls)
               15ms - ???
            0   - Semantics::analyze_impl (1 calls)
            0   - crate_def_map:wait (2 calls)
            0   - descend_into_macros (1 calls)
     1953ms - complete_expr_path
         1307ms - render_resolution
             1307ms - render_fn
                 1306ms - lang_item_query
                     1301ms - crate_lang_items_query
                            8ms - collect_lang_item
                           18ms - collect_lang_item
                           18ms - collect_lang_item
                         1222ms - collect_lang_item (344777 calls)
                            0   - crate_def_map:wait (1 calls)
                           34ms - ???
                        5ms - ???
                    0   - crate_def_map:wait (7 calls)
           27ms - PerNs::filter_visibility (235244 calls)
            1ms - crate_def_map:wait (2049 calls)
            7ms - generic_params_query (1681 calls)
            0   - item::Builder::build (2048 calls)
           64ms - render_resolution (2047 calls)
          544ms - ???
        0   - Semantics::analyze_impl (1 calls)
        0   - crate_def_map:wait (2 calls)
       16ms - ???
```
The time spent in `render_resolution` is greatly reduced and results show up in the IDE right after the trace is printed to the console. The compromise here is that only 2048 items are displayed at most (but the relevant ones are likely at the top of the list).