@Veykril 2022-08-31
I'm a bit confused here, are you actually `*` (glob) importing the bindgen file into your files? Or are you editing inside the bindgen file? Otherwise you shouldn't have these problems, and doing either of that sounds like a bad choice imo. I'd rather not impose a limit on completions generations here personally (also because this change is rather invasive, given we have to touch up every "add" function here)