@tech-ticks 2022-08-31
> I'm a bit confused here, are you actually `*` (glob) importing the bindgen file into your files? Or are you editing inside the bindgen file? Otherwise you shouldn't have these problems, and doing either of that sounds like a bad choice imo. I'd rather not impose a limit on completions generations here personally (also because this change is rather invasive, given we have to touch up every "add" function here)

@Veykril Sorry, I should have gone into more detail in the description... It's unfortunately not only an issue with glob imports. rust-analyzer also hangs when you type `crate_name::...`, which is unfortunate because the actual query behind the autocomplete still executes in a reasonable time, it's just the conversion in `render_resolution` and I/O or processing in VS Code that take way too long.

The major issue here is that even if you type e.g. `crate_name::prefix_`, rust-analyzer will still render and send all autocomplete items in `crate_name` without any filtering and then the fuzzy string matching is done by VS Code on the entire data set. Performance is more acceptable on a lower scale here, but completely breaks with a lot of items. I didn't notice that rust-analyzer was lacking pre-filtering here, so relevant completion items are actually missing in the current state of the PR. I wanted to tackle that too (and already have a prototype implementation that is a lot faster with this project), but didn't have time to finish it yet.

I do realize that autocompletion for a crate with 200,000 items is an extremely ridiculous usecase, but I think that at least filtering items on the rust-analyzer side instead of relying on VS Code would yield performance improvements in less extreme cases as well since it would save time on processing in `render_resolution`, serializing a large chunk of JSON data and processing it in VS Code. Limiting the total number of autocomplete items seemed like a good idea to me as well to save on some resources in cases where users probably won't look through thousands of suggestions, but there are of course some drawbacks as well as you've mentioned. Either way, the lack of pre-filtering items based on their name seems to be the main culprit for more common use cases here. Feel free to close this PR, I'd either way like to follow up on filtering sometime if you agree that it would be useful.