@flodiebold 2022-08-31
Maybe we should instead just limit the number of returned completions on the LSP layer?