@flodiebold 2022-08-31
Also, maybe a possible workaround would be to mark the bindgen-generated items `#[doc(hidden)]`? I can't imagine you can generate a useful documentation file with 200000 items either...