@matklad 2022-09-18
Couple of thoughts: 

* Yes, we should limit this server-side. That's a general rule for I think any LSP request -- there's never a need to return more items which can be displayed on a single screen
* We should limit this in the LSP layer **after** we sorted completions in the order of relevance 
* which means that https://github.com/rust-lang/rust-analyzer/issues/7935 is sadly a pre-requisite 


What we can do immediately though, is to make this configurable, and have infinite limit by default.