@matklad 2022-09-18
Partial mentoring instructions for lsp-layer limiting: https://users.rust-lang.org/t/how-many-symbols-can-rust-analyzer-autocomplete-in-practice/81400/5?u=matklad