@tech-ticks 2022-09-18
> Couple of thoughts:
> 
> * Yes, we should limit this server-side. That's a general rule for I think any LSP request -- there's never a need to return more items which can be displayed on a single screen
> * We should limit this in the LSP layer **after** we sorted completions in the order of relevance
> * which means that [Implement server-side completion sorting and filtering #7935](https://github.com/rust-lang/rust-analyzer/issues/7935) is sadly a pre-requisite

Yeah, I've already experimented with server-side filtering a bit (only with non-fuzzy matching) but I'm unfortunately a bit low on time right now. Maybe I'll have resources to look into it again in the next few weeks but I can't promise that.

> What we can do immediately though, is to make this configurable, and have infinite limit by default.

I could update this PR to make the limit configurable. I'm not sure if it actually makes sense though since it would exclude relevant items in the current state, which isn't what anyone would want, unless if it would be used in conjunction with the server-side filtering later.